[Performance] Compare RSSM higher-order scan
