+++ torchrl/modules/models/__init__.py
@@ -8,19 +8,12 @@ import sys
 
 from torchrl.modules.tensordict_module.common import DistributionalDQNnet
 
-from . import dreamer_v3
+from . import model_based
 from .act import ACTModel
 from .batchrenorm import BatchRenorm1d
 from .cross_group_critic import CrossCriticGroupSpec, CrossGroupCritic
 
 from .decision_transformer import DecisionTransformer
-from .dreamer_v3 import (
-    DreamerV3MLP,
-    RSSMPosteriorV3,
-    RSSMPriorV3,
-    RSSMRolloutV3,
-    SymExpTwoHot,
-)
 from .exploration import (
     ConsistentDropout,
     ConsistentDropoutModule,
@@ -32,11 +25,16 @@ from .gp import GPWorldModel
 from .llm import GPT2RewardModel
 from .model_based import (
     DreamerActor,
+    DreamerV3MLP,
     ObsDecoder,
     ObsEncoder,
     RSSMPosterior,
+    RSSMPosteriorV3,
     RSSMPrior,
+    RSSMPriorV3,
     RSSMRollout,
+    RSSMRolloutV3,
+    SymExpTwoHot,
 )
 from .models import (
     Conv2dNet,
@@ -63,9 +61,10 @@ from .rbf_controller import RBFController
 from .utils import Squeeze2dLayer, SqueezeLayer
 
 
-# Preserve the module path shipped in TorchRL 0.13 without keeping a duplicate
-# compatibility file. New code should import from ``dreamer_v3``.
-sys.modules[f"{__name__}.model_based_v3"] = dreamer_v3
+# Preserve the module paths shipped in TorchRL 0.13 without keeping duplicate
+# compatibility files. New code should import from ``model_based``.
+sys.modules[f"{__name__}.dreamer_v3"] = model_based
+sys.modules[f"{__name__}.model_based_v3"] = model_based
 
 __all__ = [
     "ACTModel",
